refactor(elements): extract Snapshot sub-module
Phase 12a §5 of the CreateElementMediator god-module split.
Moves the post-race element snapshotting + identity capture
cluster out of CreateElementMediator.ts into Create/Snapshot.ts.

Why
---
After §4 (Hittest+Entries) the parent still owned the entire
"winner found, now capture its identity + value" cluster: stale-
element-safe value snapshot, browser-side identity extraction,
verbose payload normalisation, log-payload shaping, and the
`buildFoundResult` packaging step every resolver calls on success.
That code is pure post-race data marshalling — no business
semantics, no race plumbing, no resolver orchestration. Pulling
it out clears 266 lines from the parent (now 1537 LoC, target
≤150) and isolates the browser-side `evaluate(...)` callbacks
from the rest of the mediator so they cannot accidentally close
over node-context references.

What
----
- Create/Snapshot.ts (new) — `snapshotValue`,
  `extractAndTraceIdentity`, `buildFoundResult`, `IWinnerInfo`.
  Eleven private helpers stay encapsulated (browser-side eval
  callbacks `walkUpToAnchorHref`/`traceElementInfo`/
  `snapshotIdentityInBrowser`, partial-payload normalisers
  `resolveOuterHtml`/`normalizeVerbose`, log-payload shaper
  `buildIdentityLogPayload`, sentinels `NO_ATTR`/`UNKNOWN_IDENTITY`/
  `UNKNOWN_VERBOSE`/`OUTER_HTML_SNIPPET_MAX`, internal-only
  `IIdentityVerbose` type, trace emitters `traceSnapshot`/
  `traceElementIdentity`, href resolver `resolveHrefSnapshot`,
  full-shape capturer `extractIdentityVerbose`).
- Create/index.ts — barrel re-exports the four new public symbols.
- CreateElementMediator.ts — drops the extracted 266-line block;
  pulls the four symbols from the barrel; drops the now-unused
  `maskVisibleText` import (its only call-sites moved into
  Snapshot). `formatLocatorDetail` stays in the parent for this
  commit — it returns `string`, so exporting it would trip
  architecture Rule #15. The Race cluster (Phase 12a §6) owns
  diagnostic formatting outright, so it'll move there together
  with `traceRaceDiagnostic`.
- .github/docs-coverage-allowlist.txt — four allowlist entries
  (Reason per symbol) for the new internal exports.

Public surface
--------------
lib/index.d.ts + lib/index.d.cts SHA256 BYTE-IDENTICAL to baseline
(both 14823 B, hash 221E1611…AAB055).
lib/index.cjs grew by 491 B from a single init wrapper for the
new file — implementation reorganisation, not API change.

Verification
------------
- tsc --noEmit: clean
- eslint Mediator/Elements/Create + parent: clean
- prettier: clean
- lint:architecture (src/Scrapers/Pipeline): 0 violations
- jest Mediator/Elements (30 suites / 270 tests): all pass
- npm run build: success; d.ts hashes match baseline
