+++ src/Scrapers/Pipeline/Mediator/Elements/CreateElementMediator.ts
@@ -12,7 +12,6 @@ import { WK_DASHBOARD } from '../../Registry/WK/DashboardWK.js';
 import { WK_LOGIN_FORM } from '../../Registry/WK/LoginWK.js';
 import { capTimeout, getDebug } from '../../Types/Debug.js';
 import { toErrorMessage } from '../../Types/ErrorUtils.js';
-import { maskVisibleText } from '../../Types/LogEvent.js';
 import { none, type Option, some } from '../../Types/Option.js';
 import { fail, isOk, type Procedure, succeed } from '../../Types/Procedure.js';
 import { discoverFormAnchor, type IFormAnchor, scopeCandidates } from '../Form/FormAnchor.js';
@@ -35,9 +34,11 @@ import {
 } from './ActionExecutors.js';
 import {
   buildCandidateLocators,
+  buildFoundResult,
   buildLocatorEntries,
   buildLocatorEntriesAll,
   CLICK_RACE_TIMEOUT,
+  extractAndTraceIdentity,
   getActivePhase,
   getActiveStage,
   type ILocatorEntry,
@@ -47,6 +48,7 @@ import {
   raceLocatorsWithHitTest,
   setActivePhase as setGlobalPhase,
   setActiveStage as setGlobalStage,
+  snapshotValue,
 } from './Create/index.js';
 import {
   type IActionMediator,
@@ -278,274 +280,6 @@ function buildScopeToForm(cache: IFormCache): IElementMediator['scopeToForm'] {
     const mutable = [...candidates];
     return scopeCandidates(cache.selector, mutable);
   };
-}
-
-/**
- * Walk up DOM from element to nearest `<a>` ancestor and return its href.
- * Structural CSS for extraction — allowed per CLAUDE.md exceptions.
- * Uses `closest('a')` for flat, null-safe ancestor traversal.
- * @param el - Starting DOM element.
- * @returns href string from the nearest anchor ancestor, or empty string.
- */
-function walkUpToAnchorHref(el: Element): string {
-  const anchor = el.closest('a');
-  if (anchor) return anchor.href;
-  return '';
-}
-
-/**
- * Snapshot the element value immediately to prevent stale-element errors.
- * For target:'href' candidates, captures the href attribute.
- * Walks up to nearest `<a>` ancestor if element itself has no href.
- * Otherwise captures innerText.
- * @param entry - The winning locator entry.
- * @returns The captured text or href value.
- */
-/**
- * Extract diagnostic info from a DOM element for trace logging.
- * @param el - DOM element to inspect.
- * @returns Formatted diagnostic string.
- */
-/** Sentinel for absent attribute values. */
-const NO_ATTR = '(none)';
-
-/**
- * Extract diagnostic trace info from a DOM element.
- * Inlined ternaries to fit under the 10-LoC cap.
- * @param el - The DOM element.
- * @returns Diagnostic string with tag, text, href, aria.
- */
-function traceElementInfo(el: Element): string {
-  const rawText = el.textContent;
-  const text = rawText ? rawText.slice(0, 30).trim() : NO_ATTR;
-  const href = el.getAttribute('href') ?? NO_ATTR;
-  const aria = el.getAttribute('aria-label') ?? NO_ATTR;
-  const closestA = el.closest('a');
-  const aHref = closestA ? (closestA.getAttribute('href') ?? NO_ATTR) : 'NO_ANCHOR';
-  return `tag=${el.tagName} text=${text} href=${href} aria=${aria} closestA=${aHref}`;
-}
-
-/**
- * Emit the snapshotValue debug trace.
- * @param elInfo - Result of traceElementInfo (or 'error').
- * @param candidateInfo - Candidate kind=value descriptor.
- * @returns True after logging.
- */
-function traceSnapshot(elInfo: string, candidateInfo: string): true {
-  const elMasked = maskVisibleText(elInfo);
-  const candMasked = maskVisibleText(candidateInfo);
-  LOG.debug({ message: `snapshotValue: [${elMasked}] candidate=${candMasked}` });
-  return true;
-}
-
-/**
- * Resolve href snapshot: try directly, then walk up to nearest ancestor anchor.
- * @param entry - The winning locator entry.
- * @returns Resolved href or empty string.
- */
-async function resolveHrefSnapshot(entry: ILocatorEntry): Promise<string> {
-  const directHref = await entry.locator.getAttribute('href').catch((): string => '');
-  if (directHref) return directHref;
-  return entry.locator.evaluate(walkUpToAnchorHref).catch((): string => '');
-}
-
-/**
- * Snapshot the element value immediately to prevent stale-element errors.
- * For target:'href' candidates, captures href + walks up to nearest `<a>` ancestor.
- * Otherwise captures innerText.
- * @param entry - The winning locator entry.
- * @returns The captured text or href value.
- */
-async function snapshotValue(entry: ILocatorEntry): Promise<string> {
-  const target = entry.candidate.target ?? 'self';
-  if (target !== 'href') return entry.locator.innerText().catch((): string => '');
-  const elInfo = await entry.locator.evaluate(traceElementInfo).catch((): string => 'error');
-  traceSnapshot(elInfo, `${entry.candidate.kind}="${entry.candidate.value}"`);
-  return resolveHrefSnapshot(entry);
-}
-
-/** Post-race winner details bundled to satisfy the 3-param ceiling. */
-interface IWinnerInfo {
-  readonly index: number;
-  readonly value: string;
-  readonly identity: IElementIdentity;
-}
-
-/**
- * Build a successful IRaceResult from a winning entry.
- * @param entry - The winning locator entry.
- * @param winner - Winner index + snapshot value + DOM identity captured at
- *   resolve time (used by ACTION to build a precise click selector).
- * @returns A found IRaceResult.
- */
-function buildFoundResult(entry: ILocatorEntry, winner: IWinnerInfo): IRaceResult {
-  const { locator, candidate, context } = entry;
-  const { index, value, identity } = winner;
-  return { found: true, locator, candidate, context, index, value, identity };
-}
-
-/** Identity for "?" — used when evaluate() throws. */
-const UNKNOWN_IDENTITY: IElementIdentity = {
-  tag: '?',
-  id: '?',
-  classes: '?',
-  name: '?',
-  type: '?',
-  ariaLabel: '?',
-  title: '?',
-  href: '?',
-};
-
-/** Max chars of outerHTML to surface in trace logs (forensic snippet). */
-const OUTER_HTML_SNIPPET_MAX = 300;
-
-/** Identity bundled with a bounded outerHTML snippet — single-evaluate result. */
-interface IIdentityVerbose {
-  readonly identity: IElementIdentity;
-  readonly outerHtml: string;
-}
-
-/** Fallback verbose payload — used when evaluate throws. */
-const UNKNOWN_VERBOSE: IIdentityVerbose = { identity: UNKNOWN_IDENTITY, outerHtml: '?' };
-
-/**
- * Extract Physical Identity, log at TRACE level, return identity for the
- * caller (so ACTION-stage selectors can use the resolved element's actual
- * attributes — id/name/aria-label/title/href — rather than re-deriving from
- * the original WK candidate). The trace emit uses `domId` (not `id`) so the
- * DOM identity bypasses the credential-id redaction path; DOM ids on a public
- * commercial site are not PII. An outerHTML snippet (≤300 chars) is also
- * surfaced so CI-vs-LOCAL diffs can prove same-element-clicked.
- * @param entry - The winning locator entry.
- * @returns Identity object captured during PRE.
- */
-/**
- * Normalize a partial verbose payload to a fully-defined IIdentityVerbose,
- * substituting the UNKNOWN sentinel when the shape is wrong. Defensive
- * against test mocks that return `false`/`true`/strings from `evaluate(...)`
- * rather than the structured payload.
- * @param obj - Partial payload — typed (no `unknown`).
- * @returns Verbose shape, never with missing fields.
- */
-
-/**
- * Resolve `obj.outerHtml` to a string when present, falling back to `?`
- * when the field is missing or non-string (test mocks may return non-string
- * payloads).
- * @param obj - Partial verbose payload.
- * @returns A guaranteed outerHTML snippet.
- */
-function resolveOuterHtml(obj: Partial<IIdentityVerbose>): string {
-  if (typeof obj.outerHtml === 'string') return obj.outerHtml;
-  return '?';
-}
-
-/**
- * Normalize a partial verbose payload to a fully-defined `IIdentityVerbose`,
- * substituting the UNKNOWN sentinels when the shape is wrong. Defensive
- * against test mocks that return `false`/`true`/strings from `evaluate(...)`
- * rather than the structured payload.
- * @param obj - Partial payload from evaluate.
- * @returns Verbose shape with all fields defined.
- */
-function normalizeVerbose(obj: Partial<IIdentityVerbose>): IIdentityVerbose {
-  const identity = obj.identity ?? UNKNOWN_IDENTITY;
-  const outerHtml = resolveOuterHtml(obj);
-  return { identity, outerHtml };
-}
-
-/**
- * Capture the resolved element's identity payload (browser-side eval).
- * Must be a top-level pure function (no captured closures) so Playwright's
- * `evaluate(...)` serialization can transport it. The caller body stays
- * ≤10 LoC by delegating attribute extraction and outerHTML truncation to
- * this helper.
- * @param el - The DOM element under inspection (browser context).
- * @param max - Max length for the outerHTML snippet.
- * @returns Verbose identity payload (identity bundle + bounded outerHTML).
- */
-function snapshotIdentityInBrowser(el: Element, max: number): IIdentityVerbose {
-  const identity = {
-    tag: el.tagName,
-    id: el.id || '(none)',
-    classes: el.className || '(none)',
-    name: el.getAttribute('name') ?? '(none)',
-    type: el.getAttribute('type') ?? '(none)',
-    ariaLabel: el.getAttribute('aria-label') ?? '(none)',
-    title: el.getAttribute('title') ?? '(none)',
-    href: el.getAttribute('href') ?? '(none)',
-  };
-  const outerHtml = (el.outerHTML || '').slice(0, max);
-  return { identity, outerHtml };
-}
-
-/**
- * Run the browser-side identity snapshot for an entry and normalise the
- * payload against malformed test mocks (which may return `false` / strings
- * instead of the structured shape).
- * @param entry - The winning locator entry.
- * @returns Verbose payload (never with missing fields).
- */
-async function extractIdentityVerbose(entry: ILocatorEntry): Promise<IIdentityVerbose> {
-  const evaluated = await entry.locator
-    .evaluate(snapshotIdentityInBrowser, OUTER_HTML_SNIPPET_MAX)
-    .catch((): IIdentityVerbose => UNKNOWN_VERBOSE);
-  return normalizeVerbose(evaluated);
-}
-
-/**
- * Build the LOG.debug payload from the identity bundle. Top-level so the
- * caller (`traceElementIdentity`) stays under cap by delegating both the
- * sub-attrs projection and the outer-shape composition here.
- * @param identity - Element identity bundle.
- * @param outerHtml - Bounded outerHTML snippet.
- * @returns Pino-shaped object ready for LOG.debug.
- */
-function buildIdentityLogPayload(identity: IElementIdentity, outerHtml: string): object {
-  const attrs = {
-    name: identity.name,
-    type: identity.type,
-    ariaLabel: identity.ariaLabel,
-    title: identity.title,
-    href: identity.href,
-  };
-  return {
-    tag: identity.tag,
-    domId: identity.id,
-    classes: identity.classes,
-    attrs,
-    outerHtml,
-    visibility: 'visible',
-  };
-}
-
-/**
- * Emit the DOM identity bundle at debug level. The `domId` key (not `id`)
- * bypasses the credential-id Pino redaction — DOM ids on a public commercial
- * site are not PII.
- * @param identity - The element identity bundle.
- * @param outerHtml - Bounded outerHTML snippet from the same element.
- * @returns Sentinel ``true`` once the log has been emitted.
- */
-function traceElementIdentity(identity: IElementIdentity, outerHtml: string): true {
-  const payload = buildIdentityLogPayload(identity, outerHtml);
-  LOG.debug(payload);
-  return true;
-}
-
-/**
- * Snapshot the resolved element's DOM identity (tag, id, classes, attrs)
- * plus a bounded outerHTML, log the bundle at debug level, and return the
- * identity for ACTION-stage selector building. The trace emit uses `domId`
- * (not `id`) so the public DOM identity bypasses the credential-id Pino
- * redaction — DOM ids on a public commercial site are not PII.
- * @param entry - The winning locator entry.
- * @returns Identity object captured during PRE.
- */
-async function extractAndTraceIdentity(entry: ILocatorEntry): Promise<IElementIdentity> {
-  const { identity, outerHtml } = await extractIdentityVerbose(entry);
-  traceElementIdentity(identity, outerHtml);
-  return identity;
 }
 
 /**
